refactor: mv code from init to manager.py
